# api/openapi.yaml
openapi: 3.0.0
info:
  title: "ONDC Protocol API for retail (grocery, f&b)"
  description: "ONDC Protocol API specification, which includes statutory requirements\
    \ for packaged commodities and pre-packaged food This is an adaptation of Beckn\
    \ Core version 0.9.3"
  version: 1.0.22
...
servers:
- url: https://localhost:8080/
  description: SwaggerHub API Auto Mocking
security:
- SubscriberAuth: []
- GatewaySubscriberAuth: []
- GatewaySubscriberAuthNew: []
paths:
  /search:
    post:
      tags:
      - ONDC Seller App
      - ONDC Gateway
      description: Search for services by intent
      operationId: searchPOST
      requestBody:
        description: Buyer searches for products and services
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/search_body'
      responses:
        "200":
          description: Acknowledgement of message received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
      x-swagger-router-controller: ONDCSellerApp
  /select:
    post:
      tags:
      - ONDC Seller App
      description: Select items from the catalog and build your order
      operationId: selectPOST
      requestBody:
        description: Buyer selects one or more catalog items
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/select_body'
      responses:
        "200":
          description: Acknowledgement of message received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
      x-swagger-router-controller: ONDCSellerApp
...
  /on_search:
    post:
      tags:
      - ONDC Buyer App
      - ONDC Gateway
      description: Send catalog
      operationId: on_searchPOST
      requestBody:
        description: Sellers provide their catalog in response to buyer search
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/on_search_body'
      responses:
        "200":
          description: Acknowledgement of message received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
      x-swagger-router-controller: ONDCBuyerApp
  /on_select:
    post:
      tags:
      - ONDC Buyer App
      description: Send draft order object with quoted price for selected items
      operationId: on_selectPOST
      requestBody:
        description: Seller provides quote for selected items
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/on_select_body'
      responses:
        "200":
          description: Acknowledgement of message received
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
      x-swagger-router-controller: ONDCBuyerApp
...
components:
  schemas:
    Ack:
      required:
      - status
      type: object
      properties:
        status:
          type: string
          description: "Describe the status of the ACK response. If schema validation\
            \ passes, status is ACK else it is NACK"
          enum:
          - ACK
          - NACK
      description: Describes the ACK response
      example:
        status: ACK
...
    Catalog:
      required:
      - bpp/descriptor
      - bpp/providers
      type: object
      properties:
        bpp/descriptor:
          $ref: '#/components/schemas/Descriptor'
        bpp/categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        bpp/fulfillments:
          type: array
          items:
            $ref: '#/components/schemas/Fulfillment'
        bpp/payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        bpp/offers:
          type: array
          items:
            $ref: '#/components/schemas/Offer'
        bpp/providers:
          required:
          - Provider.fulfillments.id
          - Provider.fulfillments.type
          - Provider.locations.address.area_code
          - Provider.locations.address.city
          - Provider.locations.address.state
          - Provider.locations.address.street
          - Provider.locations.gps
          - Provider.locations.id
          type: array
          items:
            $ref: '#/components/schemas/Provider'
        exp:
          type: string
          description: Time after which catalog has to be refreshed
          format: date-time
      description: Describes a Seller App catalog
...
    Context:
      required:
      - action
      - bap_id
      - bap_uri
      - city
      - core_version
      - country
      - domain
      - message_id
      - timestamp
      - transaction_id
      type: object
      properties:
        domain:
          $ref: '#/components/schemas/Domain'
        country:
          $ref: '#/components/schemas/Country/properties/code'
        city:
          $ref: '#/components/schemas/City/properties/code'
        action:
          type: string
          description: Defines the ONDC API call. Any actions other than the enumerated
            actions are not supported by ONDC Protocol
          enum:
          - search
          - select
          - init
          - confirm
          - update
          - status
          - track
          - cancel
          - rating
          - support
          - on_search
          - on_select
          - on_init
          - on_confirm
          - on_update
          - on_status
          - on_track
          - on_cancel
          - on_rating
          - on_support
        core_version:
          type: string
          description: Version of ONDC core API specification being used
        bap_id:
          type: string
          description: Unique id of the Buyer App. By default it is the fully qualified
            domain name of the Buyer App
        bap_uri:
          type: string
          description: URI of the Buyer App for accepting callbacks. Must have the
            same domain name as the bap_id
          format: uri
        bpp_id:
          type: string
          description: "Unique id of the Seller App. By default it is the fully qualified\
            \ domain name of the Seller App, mandatory for all peer-to-peer API requests,\
            \ i.e. except search and on_search"
        bpp_uri:
          type: string
          description: "URI of the Seller App. Must have the same domain name as the\
            \ bap_id, mandatory for all peer-to-peer API requests, i.e. except search\
            \ and on_search"
          format: uri
        transaction_id:
          type: string
          description: This is a unique value which persists across all API calls
            from search through confirm
        message_id:
          type: string
          description: This is a unique value which persists during a request / callback
            cycle
        timestamp:
          type: string
          description: Time of request generation in RFC3339 format
          format: date-time
        key:
          type: string
          description: The encryption public key of the sender
        ttl:
          type: string
          description: "Timestamp for which this message holds valid in ISO8601 durations\
            \ format - Outer limit for ttl for search, select, init, confirm, status,\
            \ track, cancel, update, rating, support is 'PT30S' which is 30 seconds,\
            \ different buyer apps can change this to meet their UX requirements,\
            \ but it shouldn't exceed this outer limit"
      description: Describes a ONDC message context
...
    Descriptor:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        symbol:
          type: string
        short_desc:
          type: string
        long_desc:
          type: string
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'
        audio:
          type: string
          format: uri
        "3d_render":
          type: string
          format: uri
      description: Describes the description of a real-world object.
...
    Domain:
      type: string
      description: Codification of domain for ONDC
      enum:
      - nic2004:52110
...
    Error:
      required:
      - code
      - type
      type: object
      properties:
        type:
          type: string
          enum:
          - CONTEXT-ERROR
          - CORE-ERROR
          - DOMAIN-ERROR
          - POLICY-ERROR
          - JSON-SCHEMA-ERROR
        code:
          type: string
          description: "ONDC specific error code. For full list of error codes, refer\
            \ to docs/drafts/Error Codes.md of this repo"
        path:
          type: string
          description: Path to json schema generating the error. Used only during
            json schema validation errors
        message:
          type: string
          description: Human readable message describing the error
      description: Describes an error object
      example:
        path: path
        code: code
        type: CONTEXT-ERROR
        message: message
...
    Fulfillment:
      required:
      - id
      - type
      type: object
      properties:
        id:
          type: string
          description: Unique reference ID to the fulfillment of an order
        type:
          type: string
          description: This describes the type of fulfillment ("Pickup" - Buyer picks
            up from store by themselves or through their logistics provider; "Delivery"
            - seller delivers to buyer)
          enum:
          - Delivery
          - Pickup
          - Delivery and Pickup
          - Reverse QC
        '@ondc/org/category':
          type: string
          description: Fulfillment Category
        '@ondc/org/TAT':
          type: string
          description: fulfillment turnaround time in ISO8601 durations format e.g.
            'PT24H' indicates 24 hour TAT
        provider_id:
          $ref: '#/components/schemas/Provider/properties/id'
        '@ondc/org/provider_name':
          type: string
        rating:
          $ref: '#/components/schemas/Rating/properties/value'
        state:
          $ref: '#/components/schemas/State'
        tracking:
          type: boolean
          description: Indicates whether the fulfillment allows tracking
          default: false
        customer:
          $ref: '#/components/schemas/Fulfillment_customer'
        agent:
          $ref: '#/components/schemas/Agent'
        person:
          $ref: '#/components/schemas/Person'
        contact:
          $ref: '#/components/schemas/Contact'
        vehicle:
          $ref: '#/components/schemas/Vehicle'
        start:
          $ref: '#/components/schemas/Fulfillment_start'
        end:
          $ref: '#/components/schemas/Fulfillment_end'
        rateable:
          $ref: '#/components/schemas/Rateable'
        tags:
          $ref: '#/components/schemas/Tags'
      description: "Describes how a single product/service will be rendered/fulfilled\
        \ to the end customer - State can have following values - \"Serviceable\"\
        , \"Non-serviceable\", \"At pickup location\", \"Order picked up\", \"At drop\
        \ location\", \"Order delivered\", \"Cancelled\""
...
    Intent:
      type: object
      properties:
        descriptor:
          $ref: '#/components/schemas/Descriptor'
        provider:
          $ref: '#/components/schemas/Provider'
        fulfillment:
          $ref: '#/components/schemas/Fulfillment'
        payment:
          $ref: '#/components/schemas/Payment'
        category:
          $ref: '#/components/schemas/Category'
        offer:
          $ref: '#/components/schemas/Offer'
        item:
          $ref: '#/components/schemas/Item'
        tags:
          $ref: '#/components/schemas/Tags'
      description: Intent of a user. Used for searching for services. Buyer App can
        set finder fee type in payment."@ondc/org/buyer_app_finder_fee_type" and amount
        in "@ondc/org/buyer_app_finder_fee_amount"
...
    Order:
      type: object
      properties:
        id:
          type: string
          description: "Unique identifier for Order across the network, Will be created\
            \ by buyer app in confirm API"
        state:
          type: string
          enum:
          - Created
          - Packed
          - Shipped
          - Out for Delivery
          - Delivered
          - RTO initiated
          - RTO delivered
          - Cancelled
        provider:
          $ref: '#/components/schemas/Order_provider'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Order_items'
        add_ons:
          type: array
          items:
            $ref: '#/components/schemas/Order_add_ons'
        offers:
          type: array
          items:
            $ref: '#/components/schemas/Order_offers'
        documents:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        billing:
          $ref: '#/components/schemas/Billing'
        fulfillments:
          type: array
          items:
            $ref: '#/components/schemas/Fulfillment'
        quote:
          $ref: '#/components/schemas/Quotation'
        payment:
          $ref: '#/components/schemas/Payment'
        '@ondc/org/cancellation':
          $ref: '#/components/schemas/Cancellation'
        '@ondc/org/linked_orders':
          type: array
          description: payload for cascaded orders e.g. order for logistics services
            linked to a retail order
          items:
            $ref: '#/components/schemas/Order_ondcorglinked_orders'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      description: "Describes the details of an order - For \"@ondc/org/cancellation.cancelled_by\"\
        \ use following values - buyer-app, seller-app, logistics-provider;"
...
    Payment:
      type: object
      properties:
        uri:
          type: string
          description: "A payment uri to be called by the Buyer App. If empty, then\
            \ the payment is to be done offline. The details of payment should be\
            \ present in the params object. If ```tl_method``` = http/get, then the\
            \ payment details will be sent as url params. Two url param values, ```$transaction_id```\
            \ and ```$amount``` are mandatory. And example url would be : https://www.example.com/pay?txid=$transaction_id&amount=$amount&vpa=upiid&payee=shopez&billno=1234"
          format: uri
        tl_method:
          type: string
          enum:
          - http/get
          - http/post
          - payto
          - upi
        params:
          $ref: '#/components/schemas/Payment_params'
        type:
          type: string
          enum:
          - ON-ORDER
          - PRE-FULFILLMENT
          - ON-FULFILLMENT
          - POST-FULFILLMENT
        status:
          type: string
          enum:
          - PAID
          - NOT-PAID
          - PENDING
        time:
          $ref: '#/components/schemas/Time'
        collected_by:
          type: string
          enum:
          - BAP
          - BPP
        '@ondc/org/collected_by_status':
          type: string
          enum:
          - Assert
          - Agree
          - Disagree
          - Terminate
        '@ondc/org/buyer_app_finder_fee_type':
          type: string
          enum:
          - Amount
          - Percent
        '@ondc/org/buyer_app_finder_fee_amount':
          $ref: '#/components/schemas/DecimalValue'
        '@ondc/org/withholding_amount':
          $ref: '#/components/schemas/DecimalValue'
        '@ondc/org/withholding_amount_status':
          type: string
          enum:
          - Assert
          - Agree
          - Disagree
          - Terminate
        '@ondc/org/return_window':
          type: string
          description: return window for the item in ISO8601 durations format e.g.
            'PT24H' indicates 24 hour return window
        '@ondc/org/return_window_status':
          type: string
          enum:
          - Assert
          - Agree
          - Disagree
          - Terminate
        '@ondc/org/settlement_basis':
          type: string
          description: "In case of prepaid payment, whether settlement between counterparties\
            \ should be on the basis of collection, shipment or delivery"
          enum:
          - Collection
          - Shipment
          - Delivery
        '@ondc/org/settlement_basis_status':
          type: string
          enum:
          - Assert
          - Agree
          - Disagree
          - Terminate
        '@ondc/org/settlement_window':
          type: string
          description: return window for the item in ISO8601 durations format e.g.
            'PT24H' indicates 24 hour return window
        '@ondc/org/settlement_window_status':
          type: string
          enum:
          - Assert
          - Agree
          - Disagree
          - Terminate
        '@ondc/org/settlement_details':
          type: array
          items:
            $ref: '#/components/schemas/Payment_ondcorgsettlement_details'
      description: Describes a payment
...
    Provider:
      required:
      - '@ondc/org/fssai_license_no'
      - descriptor
      - id
      - ttl
      type: object
      properties:
        id:
          type: string
          description: Id of the provider
        descriptor:
          $ref: '#/components/schemas/Descriptor'
        category_id:
          type: string
          description: Category Id of the provider
        '@ondc/org/fssai_license_no':
          type: string
          description: FSSAI license no. Mandatory for category_id "F&B"
        rating:
          $ref: '#/components/schemas/Rating/properties/value'
        time:
          $ref: '#/components/schemas/Time'
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        fulfillments:
          type: array
          items:
            $ref: '#/components/schemas/Fulfillment'
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        locations:
          type: array
          items:
            allOf:
            - $ref: '#/components/schemas/Location'
            - type: object
              properties:
                rateable:
                  $ref: '#/components/schemas/Rateable'
        offers:
          type: array
          items:
            $ref: '#/components/schemas/Offer'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        ttl:
          type: string
          description: Validity of catalog in ISO8601 durations format after which
            it has to be refreshed e.g. 'P7D' indicates validity of 7 days; value
            of 0 indicates catalog is not cacheable
        exp:
          type: string
          description: Time after which catalog has to be refreshed
          format: date-time
        rateable:
          $ref: '#/components/schemas/Rateable'
        tags:
          $ref: '#/components/schemas/Tags'
      description: "Describes a service provider. This can be a restaurant, a hospital,\
        \ a Store etc"
...
    Tags:
      type: object
      additionalProperties:
        type: string
      description: Describes a tag. This is a simple key-value store which is used
        to contain extended metadata
...
    search_message:
      type: object
      properties:
        intent:
          $ref: '#/components/schemas/Intent'
    search_body:
      required:
      - context
      - message
      type: object
      properties:
        context:
          $ref: '#/components/schemas/Context'
        message:
          $ref: '#/components/schemas/search_message'
    inline_response_200:
      required:
      - message
      type: object
      properties:
        message:
          $ref: '#/components/schemas/inline_response_200_message'
        error:
          $ref: '#/components/schemas/Error'
      example:
        message:
          ack:
            status: ACK
        error:
          path: path
          code: code
          type: CONTEXT-ERROR
          message: message
    select_message:
      required:
      - order
      type: object
      properties:
        order:
          $ref: '#/components/schemas/Order'
    select_body:
      required:
      - context
      - message
      type: object
      properties:
        context:
          $ref: '#/components/schemas/Context'
        message:
          $ref: '#/components/schemas/select_message'
...
    on_search_message:
      required:
      - catalog
      type: object
      properties:
        catalog:
          $ref: '#/components/schemas/Catalog'
    on_search_body:
      required:
      - context
      type: object
      properties:
        context:
          $ref: '#/components/schemas/Context'
        message:
          $ref: '#/components/schemas/on_search_message'
        error:
          $ref: '#/components/schemas/Error'
    inline_response_200_1:
      required:
      - context
      type: object
      properties:
        message:
          $ref: '#/components/schemas/inline_response_200_message'
        error:
          $ref: '#/components/schemas/Error'
      example:
        message:
          ack:
            status: ACK
        error:
          path: path
          code: code
          type: CONTEXT-ERROR
          message: message
    on_select_message_order:
      type: object
      properties:
        provider:
          $ref: '#/components/schemas/Provider'
        provider_location:
          $ref: '#/components/schemas/Location'
        items:
          type: array
          items:
            allOf:
            - $ref: '#/components/schemas/Item'
            - properties:
                quantity:
                  $ref: '#/components/schemas/ItemQuantity'
        add_ons:
          type: array
          items:
            $ref: '#/components/schemas/AddOn'
        offers:
          type: array
          items:
            $ref: '#/components/schemas/Offer'
        quote:
          $ref: '#/components/schemas/Quotation'
    on_select_message:
      required:
      - order
      type: object
      properties:
        order:
          $ref: '#/components/schemas/on_select_message_order'
    on_select_body:
      required:
      - context
      type: object
      properties:
        context:
          $ref: '#/components/schemas/Context'
        message:
          $ref: '#/components/schemas/on_select_message'
        error:
          $ref: '#/components/schemas/Error'
...
    Order_provider:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Provider/properties/id'
        locations:
          maxItems: 1
          type: array
          items:
            $ref: '#/components/schemas/Order_provider_locations'
...
    inline_response_200_message:
      required:
      - ack
      type: object
      properties:
        ack:
          $ref: '#/components/schemas/Ack'
      example:
        ack:
          status: ACK
  securitySchemes:
    SubscriberAuth:
      type: apiKey
      description: "Signature of message body using Buyer App or Seller App subscriber's\
        \ signing public key. <br/><br/>Format:<br/><br/><code>Authorization : Signature\
        \ keyId=\"{subscriber_id}|{unique_key_id}|{algorithm}\",algorithm=\"ed25519\"\
        ,created=\"1606970629\",expires=\"1607030629\",headers=\"(created) (expires)\
        \ digest\",signature=\"Base64(signing string)\"</code>"
      name: Authorization
      in: header
    GatewaySubscriberAuth:
      type: apiKey
      description: "Signature of message body + BuyerApp/SellerApp's Authorization\
        \ header using Gateway's signing public key. Format:<br/><br/><code>Proxy-Authorization\
        \ : Signature keyId=\"{subscriber_id}|{unique_key_id}|{algorithm}\",algorithm=\"\
        ed25519\",created=\"1606970629\",expires=\"1607030629\",headers=\"(created)\
        \ (expires) digest\",signature=\"Base64(signing string)\"</code><p><b>Note:</b>This\
        \ header will be deprecated soon and will no longer be supported in future\
        \ releases. New implementors are requested to use the X-Gateway-Authorization\
        \ header. Existing implementations are requested to migrate their header to\
        \ the new header. The deprecation date will be set after discussion as per\
        \ the standard specification governance process.</p>"
      name: Proxy-Authorization
      in: header
    GatewaySubscriberAuthNew:
      type: apiKey
      description: "Signature of message body + BuyerApp/SellerApp's Authorization\
        \ header using Gateway's signing public key. Format:<br/><br/><code>X-Gateway-Authorization\
        \ : Signature keyId=\"{subscriber_id}|{unique_key_id}|{algorithm}\",algorithm=\"\
        ed25519\",created=\"1606970629\",expires=\"1607030629\",headers=\"(created)\
        \ (expires) digest\",signature=\"Base64(signing string)\"</code>"
      name: X-Gateway-Authorization
      in: header